Fixes 32 bit compile

+++ runtime/examples/C/arrayinit/arrayinit-example.c
@@ -98,7 +98,7 @@ int main(int argc, char **argv) {
     peid = PE_ID;
   }
 
-  printf("Using PEId %ld.\n", peid);
+  printf("Using PEId %u.\n", (uint32_t)peid);
 
   if (tapasco_device_num_pes(d, peid) == 0) {
     printf("No Arrayinit PE found.\n");
